Move InferenceImage to smart-tools
