+++ web_ui/packages/smart-tools/index.ts
@@ -6,4 +6,6 @@ export { default as OpenCVLoader } from './src/utils/opencv-loader';
 export { Watershed } from './src/watershed/watershed';
 export { type WatershedInstance, type WatershedPolygon } from './src/watershed/interfaces';
 
+export { InferenceImage } from './src/inference-image/inference-image';
+
 export { formatContourToPoints, approximateShape } from './src/utils/utils';
