+++ web_ui/src/webworkers/utils.ts
@@ -44,30 +44,6 @@ export const getPointsFromMat = (mat: OpenCVTypes.Mat, offset = { x: 0, y: 0 }):
     return points;
 };
 
-export const formatImageData = (CV: OpenCVTypes.cv, mat: OpenCVTypes.Mat): ImageData => {
-    const img = new CV.Mat();
-    const depth = mat.type() % 8;
-    const scale = depth <= CV.CV_8S ? 1 : depth <= CV.CV_32S ? 1 / 256 : 255;
-    const shift = depth === CV.CV_8S || depth === CV.CV_16S ? 128 : 0;
-
-    mat.convertTo(img, CV.CV_8U, scale, shift);
-
-    switch (img.type()) {
-        case CV.CV_8UC1:
-            CV.cvtColor(img, img, CV.COLOR_GRAY2RGBA);
-            break;
-        case CV.CV_8UC3:
-            CV.cvtColor(img, img, CV.COLOR_RGB2RGBA);
-            break;
-        case CV.CV_8UC4:
-            break;
-        default:
-            throw new Error('Bad number of channels (Source image must have 1, 3 or 4 channels)');
-    }
-
-    return new ImageData(new Uint8ClampedArray(img.data), img.cols, img.rows);
-};
-
 // For debugging purposes, not being used atm
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const logMat = (mat: OpenCVTypes.Mat, name: string): void => {
